+++ NEWS.md
@@ -2,6 +2,7 @@
 
 # development version
 
+- Add Cojutti 2024 dalbavancin ([doi:10.1007/s40262-024-01410-2](https://doi.org/10.1007/s40262-024-01410-2)) -- adults with staphylococcal osteoarticular infections.
 - Add Sager 2023 sotrovimab population PK and COVID-19 progression exposure-response models ([doi:10.1002/psp4.12958](https://doi.org/10.1002/psp4.12958)) -- non-hospitalized adults and adolescents with mild-to-moderate COVID-19 at high risk of progression, plus healthy volunteers.
 
 - Add Pitsiu 2023 atacicept ([doi:10.1002/psp4.12982](https://doi.org/10.1002/psp4.12982)) -- healthy volunteers and adults with systemic lupus erythematosus.